api-python/requirements-step2.txt
# Core dependencies (already installed)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2

# Document processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0
minio==7.2.0
langdetect==1.0.9

# Basic text processing
regex==2023.10.3
